inference-service/requirements.txt
flask==3.0.3
gunicorn==22.0.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
ai-edge-litert==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-auth==2.29.0
requests==2.32.2
